+++ CHANGELOG.md
@@ -22,6 +22,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - missing tab-completion and documentation for globalpoints and globaltags command
 - `party` condition now supports a location variable, to be used in a static context
 - `enitity` condition now supports variables for the entity name
+- `mmochangeclass` objective for MMOCore that listens for a player changing their class
 ### Changed
 - `spawn` event now only spawn mobs and no other entities
 ### Deprecated